// src/models/VideoBalancerOptions.ts
export interface VideoBalancerOptions {
    video_token: string;
    serial_token: string;
    serial_group: string;
    trailer_token: string;
    content_type: string;
    partner_id: number;
    domain_id: number;
    host: string;
    port: number;
    proto: string;
    ref: string;
    ref_url: string;
    ref_back: string;
    thumbnails: {
        visible: boolean;
        count: number;
        pattern: string;
    };
    user_country: string;
    adv: object;
    c_path: string;
    r_path: string;
    r_path_ua: string;
    device: {
        mobile: boolean;
        android: boolean;
        ios: boolean;
        desktop: boolean;
        browser: string;
        platform_major_version: number
    };
    lang: string;
    assets_host: string;
    player_skin: string;
    p2p: boolean;
    start_time: null;
    continue_time: null;
    subtitles: null;
    translations: null;
    movie_translations: null;
    seasons: number[];
    episodes: number[];
    season: number;
    episode: number;
    nocontrols: boolean;
    nocontrols_seasons: boolean;
    nocontrols_translations: boolean;
    nocontrols_episodes: boolean;
    nocontrols_trailer: boolean;
    nocontrols_full_video: boolean;
    poster_url: null;
    autoplay: boolean;
    autoswitch: boolean;
    next_episode_button: boolean;
    track_watching: boolean;
    env: string;
    vast_short_skip_text: boolean;
}